requirements.txt
Django==5.0.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-nested-routers==0.93.5
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
black==24.4.2
isort==5.13.2
flake8==7.0.0
